// frontend/src/openapi/models/EvaluatorCatalog.ts
export interface EvaluatorCatalog {
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly id: string;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly name: string;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly displayName: string;
    /**
     *
     * @type {number}
     * @memberof EvaluatorCatalog
     */
    readonly version: number;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly description: string;
    /**
     *
     * @type {EvaluatorKindEnum}
     * @memberof EvaluatorCatalog
     */
    readonly kind: EvaluatorKindEnum;
    /**
     *
     * @type {EvaluatorScopeEnum}
     * @memberof EvaluatorCatalog
     */
    readonly scope: EvaluatorScopeEnum;
    /**
     *
     * @type {ScoreDataTypeEnum}
     * @memberof EvaluatorCatalog
     */
    readonly scoreType: ScoreDataTypeEnum;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly capability: string | null;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly capabilityName: string;
    /**
     *
     * @type {boolean}
     * @memberof EvaluatorCatalog
     */
    readonly isGeneric: boolean;
    /**
     *
     * @type {boolean}
     * @memberof EvaluatorCatalog
     */
    readonly isManaged: boolean;
    /**
     *
     * @type {any}
     * @memberof EvaluatorCatalog
     */
    readonly provenance: any | null;
    /**
     *
     * @type {Array<string>}
     * @memberof EvaluatorCatalog
     */
    readonly applicableRoles: Array<string>;
    /**
     *
     * @type {number}
     * @memberof EvaluatorCatalog
     */
    readonly latestScore: number | null;
    /**
     *
     * @type {number}
     * @memberof EvaluatorCatalog
     */
    readonly previousScore: number | null;
    /**
     *
     * @type {number}
     * @memberof EvaluatorCatalog
     */
    readonly delta: number | null;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly latestRunId: string | null;
    /**
     *
     * @type {string}
     * @memberof EvaluatorCatalog
     */
    readonly previousRunId: string | null;
    /**
     *
     * @type {Date}
     * @memberof EvaluatorCatalog
     */
    readonly latestRunAt: Date | null;
    /**
     *
     * @type {Date}
     * @memberof EvaluatorCatalog
     */
    readonly previousRunAt: Date | null;
}